+++ dart/test/sentry_test.dart
@@ -102,6 +102,14 @@ void main() {
       expect(client.captureEventCalls.first.scope, isNotNull);
     });
 
+    test('should capture exception with message', () async {
+      await Sentry.captureException(anException,
+          message: SentryMessage('Sentry rocks'));
+
+      expect(client.captureEventCalls.first.event.message?.formatted,
+          'Sentry rocks');
+    });
+
     test('should capture exception withScope', () async {
       await Sentry.captureException(anException, withScope: (scope) {
         scope.setUser(SentryUser(id: 'foo bar'));